# docs/swagger.yaml
basePath: /
definitions:
...
  handler.HTTPError:
    properties:
      error:
        type: string
    type: object
  handler.HTTPMessage:
    properties:
      message:
        type: string
    type: object
...
  handler.MollieOrderView:
    properties:
      checkout:
        type: string
      id:
        type: string
      profileId:
        type: string
    type: object
  handler.MollieSubmitRequest:
    properties:
      language:
        type: string
      method:
        type: string
    type: object
...
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: GoShop full featured api, see documentation https://docs.google.com/document/d/1VlkAYTqZG9oGvZxnSNDpcuafU-msOb2Qm2FevMJkA7w/edit
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GoShop API
  version: "1.0"
paths:
...
  /api/v1/account/orders/{id}/mollie/submit:
    post:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: body
        in: body
        name: form
        required: true
        schema:
          $ref: '#/definitions/handler.MollieSubmitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MollieOrderView'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      summary: Post mollie order
      tags:
      - account
      - frontend
  /api/v1/account/orders/{id}/mollie/success:
    get:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.HTTPMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      summary: Get mollie payment success
      tags:
      - account
      - frontend
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"